Improved nwb spec generation - Separate nwb spec files generation and loading
